requirements.txt
# Core Streamlit Framework
streamlit>=1.28.0,<2.0.0

# Data Processing and Analysis
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Data Visualization
plotly>=5.17.0,<6.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# AI and Machine Learning
openai>=1.6.1,<2.0.0
google-generativeai>=0.3.2,<1.0.0
langchain>=0.0.350,<1.0.0
langchain-openai>=0.0.2,<1.0.0
tiktoken>=0.5.2,<0.6.0

# Environment and Configuration
python-dotenv>=1.0.0,<2.0.0

# Development and Testing (optional)
pytest>=7.4.0,<8.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0

# Production and Deployment
gunicorn>=21.0.0,<22.0.0
